Improved forms tests
Splitted non tests relevant code into different files. Added real wait of the server when the form is submitted.
